Added more buildable projects

+++ Part #7 - Mappers & Basic Sounds/olc2A03.h
@@ -217,11 +217,11 @@ private:
 			double b = 0;
 			double p = dutycycle * 2.0 * pi;
 
-			auto approxsin = [](float t)
+			auto approxsin = [](double t)
 			{
-				float j = t * 0.15915;
+				double j = t * 0.15915;
 				j = j - (int)j;
-				return 20.785 * j * (j - 0.5) * (j - 1.0f);
+				return 20.785 * j * (j - 0.5) * (j - 1.0);
 			};
 
 			for (double n = 1; n < harmonics; n++)
